chore(server): implement feedback on server
- Setup eslint with the airbnb configurations
- Return status codes in all responses
- Model responses after those specified in the rubric
- Rename api folder to server
- Update package.json to remove empty lines.
- Update readme with details sent to the channel (provide parameters required for testing endpoints).
- Remove comments on regex

[delivers #162465593]
