+++ app.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import bodyParser from 'body-parser';
-import router from './api/routes';
+import router from './server/routes';
 
 const app = express();
 
@@ -18,7 +18,7 @@ app.all('*', (req, res) => {
   res.status(404).send('<h2>Well!!! This is Embarrasing</h2><p>There are no resources here. Check the documentation here for valid routes</p>');
 });
 
-const PORT = 3000;
+const PORT = process.env.PORT || 3000;
 
 app.listen(PORT, () => {
   /* eslint-disable-next-line no-console */
