+++ package.json
@@ -4,10 +4,10 @@
     "description": "iReporter enables  any/every citizen to bring any form of corruption to the notice of appropriate authorities and the  general public. Users can also report on things that needs government intervention",
     "main": "app.js",
     "scripts": {
-        "test": "nyc mocha ./api/tests/* --require babel-core/register --exit",
+        "test": "nyc mocha ./server/tests/* --require babel-core/register --exit",
         "start": "node_modules/.bin/nodemon app.js --exec babel-node --",
-        "coverage": "nyc report --reporter=text-lcov | coveralls"
-
+        "coverage": "nyc report --reporter=text-lcov | coveralls",
+        "serve": "babel-node app.js"
     },
     "repository": {
         "type": "git",
@@ -30,12 +30,11 @@
     "dependencies": {
         "body-parser": "^1.18.3",
         "express": "^4.16.4",
-        "moment": "^2.22.2"
+        "moment": "^2.22.2",
+        "babel-cli": "^6.26.0",
+        "babel-preset-env": "^1.7.0"
     },
     "devDependencies": {
-        "babel-cli": "^6.26.0",
-        "babel-preset-env": "^1.7.0",
-        "babel-watch": "^2.0.7",
         "chai": "^4.2.0",
         "chai-http": "^4.2.0",
         "coffeescript": "^2.3.2",
