+++ server/middlewares/isValidComment.js
@@ -2,7 +2,7 @@ const isValidComment = (req, res, next) => {
   if (req.body.comment.length < 50 || req.body.comment.length > 420) {
     return res.status(400)
       .send({
-        success: 'false',
+        status: 400,
         error: 'Comment too short or too long',
       });
   }
