+++ server/middlewares/isValidType.js
@@ -4,7 +4,7 @@ const isValidType = (req, res, next) => {
   if (!acceptedType.includes(req.body.type)) {
     return res.status(400)
       .send({
-        success: 'false',
+        status: 400,
         error: 'Type not of accepted value',
       });
   }
